deployed
should continue clean the deploy config
