--- config/child.js
@@ -1,3 +1,0 @@
-console.log(process.env.HAHA)
-exports.a = 123
-exports.b = 123
