+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon --watch webpack.config.js --watch ./config/webpack.parts.js --exec \"webpack-dev-server --env development\"",
-    "build": "webpack --env production"
+    "build": "webpack --env production",
+    "deploy_test": "node ./config/deploy --dest=test"
   },
   "keywords": [
     "webpack",
@@ -25,6 +26,7 @@
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.4.0",
     "babili-webpack-plugin": "0.0.11",
+    "chalk": "^1.1.3",
     "clean-webpack-plugin": "^0.1.16",
     "css-loader": "^0.28.0",
     "cssnano": "^3.10.0",
@@ -46,6 +48,7 @@
     "less-loader": "^4.0.3",
     "nodemon": "^1.11.0",
     "optimize-css-assets-webpack-plugin": "^1.3.1",
+    "ora": "^1.2.0",
     "postcss-loader": "^1.3.3",
     "precss": "^1.4.0",
     "style-loader": "^0.16.1",
